# docs/swagger.yaml
definitions:
  models.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
paths:
  /backtest/data:
    get:
      consumes:
      - application/json
      description: 获取指定交易对的回测数据
      parameters:
      - description: 交易对符号，例如 BTCUSDT
        in: query
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 获取回测数据
      tags:
      - 回测
  /backtest/parquet:
    get:
      consumes:
      - application/json
      description: 获取指定交易对和日期范围的Parquet格式回测数据
      parameters:
      - description: 交易对符号，例如 BTCUSDT
        in: query
        name: symbol
        required: true
        type: string
      - description: 开始日期，格式：YYYY-MM-DD
        in: query
        name: start_date
        type: string
      - description: 结束日期，格式：YYYY-MM-DD
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 获取Parquet格式的回测数据
      tags:
      - 回测
  /health:
    get:
      consumes:
      - application/json
      description: 检查量化数据引擎API是否正常运行
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: 健康检查
      tags:
      - 系统
  /market/data:
    get:
      consumes:
      - application/json
      description: 获取指定交易对的市场数据
      parameters:
      - description: 交易对符号，例如 BTCUSDT
        in: query
        name: symbol
        required: true
        type: string
      - description: 返回数据条数，默认10
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 获取市场数据
      tags:
      - 市场
  /stock/fetch-list:
    post:
      consumes:
      - application/json
      description: 手动触发从 Tushare API 获取股票列表并保存到数据库中
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 手动触发获取股票列表
      tags:
      - 股票
swagger: "2.0"